// src/app/shared/models/Result.ts
export class Result {
    // 'vote_count': number;
    'id': number;
    // 'video': boolean ;
   'vote_average': number;
    'title': string;
    'popularity': number;
//    'original_language': string;
//     'original_title': string;
//     'genre_ids': [
//       number,
//       number,
//       number
//     ];
//     'backdrop_path': string;
//     'adult': boolean;
    'overview': string;
    'release_date': string ;
    'poster_path': string;
}